# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Netflix Clone API",
        "description": "API Documentation for Netflix Clone Video Platform",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/watch-history": {
            "get": {
                "tags": [
                    "Watch History"
                ],
                "summary": "Get watch history",
                "description": "Get watch history for the authenticated user or a specific profile.",
                "operationId": "88fa02208ce6f1cb3107fb0d3c46f021",
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "query",
                        "description": "Filter by profile ID (optional)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of records to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of watch history entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "video": {
                                                        "$ref": "#/components/schemas/Video"
                                                    },
                                                    "watched_seconds": {
                                                        "type": "integer"
                                                    },
                                                    "video_duration": {
                                                        "type": "integer"
                                                    },
                                                    "progress_percentage": {
                                                        "type": "integer"
                                                    },
                                                    "completed": {
                                                        "type": "boolean"
                                                    },
                                                    "last_watched_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Watch History"
                ],
                "summary": "Clear watch history",
                "description": "Clear all watch history for the user or a specific profile.",
                "operationId": "70e23bfbd8d2f6196cccc4d7350459c8",
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "query",
                        "description": "Clear history only for this profile",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watch history cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Watch history cleared successfully"
                                        },
                                        "entries_deleted": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/watch-history/continue-watching": {
            "get": {
                "tags": [
                    "Watch History"
                ],
                "summary": "Get videos that can be resumed",
                "description": "Get continue watching items (incomplete videos).",
                "operationId": "8506977eb382bde0131527d3d703725e",
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "query",
                        "description": "Filter by profile ID (optional)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of records to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of videos that can be resumed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "video": {
                                                        "$ref": "#/components/schemas/Video"
                                                    },
                                                    "watched_seconds": {
                                                        "type": "integer"
                                                    },
                                                    "progress_percentage": {
                                                        "type": "integer"
                                                    },
                                                    "last_watched_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/videos/{videoId}/playback-position": {
            "get": {
                "tags": [
                    "Watch History"
                ],
                "summary": "Get playback position for a video",
                "description": "Get playback position for a specific video.",
                "operationId": "f0ab216722a0d32c22a39655bbb4bffa",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "description": "Video ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "profile_id",
                        "in": "query",
                        "description": "Profile ID (optional)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Playback position details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "watched_seconds": {
                                            "type": "integer",
                                            "example": 360
                                        },
                                        "progress_percentage": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "can_resume": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No watch history found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/watch-history/{id}": {
            "delete": {
                "tags": [
                    "Watch History"
                ],
                "summary": "Delete a watch history entry",
                "description": "Delete a watch history entry.",
                "operationId": "a6a5acaed78fa3841ce1ecdcae36b008",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Watch history entry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watch history entry deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Watch history entry deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Watch history entry not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Video": {
                "title": "Video",
                "description": "Video model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Learn the basics of Laravel framework"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/videos/laravel-intro.mp4"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://example.com/thumbnails/laravel-intro.jpg"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "published",
                            "failed"
                        ],
                        "example": "published"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}